Move Geometry2D to Video
